requirements.txt
numpy>=1.18.5
PyYAML>=5.3.1
torch>=1.13.0
transformers>=4.26
pandas
scikit-learn>=1.0
unidecode
tqdm
nltk
csv
chardet